backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
sqlalchemy==2.0.23
alembic==1.12.1
httpx==0.25.2
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
pyjwt==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
pydantic[email]==2.5.0